requirement.txt
streamlit==1.30.0
transformers==4.37.0
torch==2.2.0
Pillow==10.2.0
pyperclip==1.8.2
